tweaks, moving out functions to own file
